changing rotate matrix solution

+++ app/src/main/java/com/gallopdevs/interviewtests/simpleprogrammer/arrays/MissingNumber.java
@@ -6,13 +6,13 @@ public class MissingNumber {
 
     // How do you find the missing number in a given integer array of 1 to 100?
     // cannot be used if more than one missing number or contains duplicates
-    public static int getMissingNumber(int[] numbers, int totalCount) {
-        int expectedSum = totalCount * ((totalCount + 1) / 2);
+    public static int getMissingNumber(int[] numbers, double totalCount) {
+        double expectedSum = totalCount * ((totalCount + 1) / 2);
         int actualSum = 0;
         for (int num : numbers) {
             actualSum += num;
         }
-        return expectedSum - actualSum;
+        return (int) (expectedSum - actualSum);
     }
     // more than one missing element
     public static void printMissingNumber(int[] numbers, int count) {
